Module minidom::element

Provides an Element type, which represents DOM nodes, and a builder to create them with.

Structs

Attrs

An iterator over the attributes of an Element.

AttrsMut

An iterator over the attributes of an Element, with the values mutable.

Children

An iterator over references to child elements of an Element.

ChildrenMut

An iterator over mutable references to child elements of an Element.

Element

A struct representing a DOM Element.

ElementBuilder

A builder for Elements.

Texts

An iterator over references to child text nodes of an Element.

TextsMut

An iterator over mutable references to child text nodes of an Element.

Functions

escape

helper function to escape a &[u8] and replace all xml special characters (<, >, &, ’, “) with their corresponding xml escaped value.

Type Definitions

Nodes

An iterator over references to all child nodes of an Element.

NodesMut

An iterator over mutable references to all child nodes of an Element.